Added the Container component and cleaned up in App.js
