+++ code/src/components/Container.js
@@ -1,0 +1,61 @@
+import React, { useState, useEffect } from 'react';
+
+import Thoughts from 'components/Thoughts';
+import ThoughtsForm from 'components/ThoughtsForm';
+
+const Container = () => {
+  const [thoughts, setThoughts] = useState([]);
+  const [loading, setLoading] = useState(false);
+  const [newThought, setNewThought] = useState('');
+
+  const fetchTasks = () => {
+    setLoading(true);
+    fetch('https://happy-thoughts-technigo.herokuapp.com/thoughts')
+      .then((res) => res.json())
+      .then((data) => setThoughts(data)) 
+      .catch((error) => console.error(error))
+      .finally(() => setLoading(false));
+  }
+
+  useEffect(() => {
+    fetchTasks();
+  }, []);
+
+  const handleNewThoughtChange = (event) => {
+    setNewThought(event.target.value)
+  }
+
+  const onFormSubmit = (event) => {
+    event.preventDefault();
+
+    const options = {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        message: newThought
+      })
+    }
+
+    fetch('https://happy-thoughts-technigo.herokuapp.com/thoughts', options)
+      .then((res) => res.json())
+      .then(() => fetchTasks())
+      .finally(() => setNewThought(''));
+  }
+
+  return (
+    <div>
+      <ThoughtsForm
+        newThought={newThought}
+        onNewThoughtChange={handleNewThoughtChange}
+        onFormSubmit={onFormSubmit} />
+      <Thoughts
+        loading={loading}
+        thoughts={thoughts}
+        setThoughts={setThoughts} />
+    </div>
+  );
+}
+
+export default Container;
